refactor: タイプslug→日本語テーブルをfeature非依存モジュールへ切り出し
